+++ model/Validation.py
@@ -1,19 +1,20 @@
 from Modelingfncs import Make_Y, delay, A_from_components, Z_continuous, Z_init
 from Topologies import linear, simple_branch
+import pandas as pd
+import matplotlib.pyplot as plt
+from pprint import pprint
 
 
-def validate_Y(conf, sim_topology):
-    # TODO: ABSTRACT parameters
-    df_simulated, ps = sim_topology()
+def validate_Y(df_simulated, ps, conf):
     """Logic for error extraction of Y_n"""
     df_simulated_departures = df_simulated.loc[:, ["departures"]]
-    print("---------- Simulated Departure Data ----------")
-    pprint(df_simulated_departures.head())
+    # print("---------- Simulated Departure Data ----------")
+    # pprint(df_simulated_departures.head())
 
     generated_departures = Make_Y(ps.packets_rec - 1, conf)
     df_generated = pd.DataFrame.from_dict(generated_departures, orient='index')
-    print("---------- Generated Departure Data ----------")
-    pprint(df_generated.head())
+    # print("---------- Generated Departure Data ----------")
+    # pprint(df_generated.head())
     errors = {}
     for i in range(0, ps.packets_rec):
         errors[i] = {}
@@ -22,19 +23,19 @@ def validate_Y(conf, sim_topology):
                 df_generated.values[i][0][j] - df_simulated_departures.values[i][0][j])
 
     df_errors = pd.DataFrame.from_dict(errors, orient='index')
-    pprint(df_errors)
     df_errors["sum"] = df_errors.sum(axis=1)
 
-    print("---------- Error In Departure Times ----------")
+    # print("---------- Error In Departure Times ----------")
     ax = df_errors.plot()
     ax.set_ylabel('Quantity of Error')
     ax.set_xlabel('Packet Number')
     plt.title("Error Between Y(n) and Simulated Traffic")
     plt.show()
+    return df_generated
 
 
-def validate_Z(conf, sim_topology, flag=False):
-    df_simulated, ps = sim_topology()
+def validate_Z(df_simulated, ps, conf, flag=False):
+
     if flag:
         generated_departures = Make_Y(ps.packets_rec - 1, conf)
         data = pd.DataFrame.from_dict(generated_departures, orient='index')
@@ -48,11 +49,17 @@ def validate_Z(conf, sim_topology, flag=False):
     zeee = Z_continuous(current_packet, ps.packets_rec - 1, conf)
     for m in zeee.keys():
         z = zeee[m].transpose()
+        print(z)
         current_index_packet = m
         errors = {current_index_packet: {}}
         breakpoint = 0
         total_pkts = 0
         for j in range(0, z.cols):
+            print("J:", j)
+            # print("z[0,j]:", z[0,j])
+            # print("data[\"departures\"][current_index_packet]:", data["departures"][current_index_packet])
+            #
+            # print("data[\"departures\"][current_index_packet][j % conf.number_of_routers]:", data["departures"][current_index_packet][j % conf.number_of_routers])
             if current_index_packet < 0:
                 errors[current_index_packet]['Router {}'.format(j % conf.number_of_routers)] = 0
                 errors[current_index_packet]['packet'] = current_index_packet
@@ -76,9 +83,11 @@ def validate_Z(conf, sim_topology, flag=False):
 
     pprint(df.head())
     print("---------- Error In Departure Times ----------")
+    df = df[df.packet >= 0]
     new = df.drop(columns=["packet"], axis=1)
     pprint(new)
-    ax = new.plot()
+
+    ax = new.plot(use_index=True)
     ax.set_ylabel('Quantity of Error')
     ax.set_xlabel('Packet Number')
     plt.title("Error Between Z(n) and Simulated Traffic")
@@ -96,4 +105,4 @@ def validate_z_against_y(conf, sim_topology):
     pprint(df_generated.head())
 
 
-validate_Y()
+# validate_Y()
